Refactor pages to utilize draw.js instead of objectDraw.js
