+++ QualityControl/public/layout/view/panels/objectTreeSidebar.js
@@ -14,7 +14,7 @@
 
 import { h } from '/js/src/index.js';
 import { spinner } from '../../../common/spinner.js';
-import { draw } from '../../../object/objectDraw.js';
+import { draw } from '../../../common/object/draw.js';
 import { iconCaretBottom, iconCaretRight, iconBarChart } from '/js/src/icons.js';
 import virtualTable from '../../../object/virtualTable.js';
 
@@ -194,7 +194,14 @@ const objectPreview = (model) => {
   const isSelected = model.object.selected;
   if (isSelected) {
     const objName = model.object.selected.name;
-    return isSelected && h('.bg-white', { style: 'height: 20em' }, draw(model, objName, {}));
+    const remoteObject = model.object.objects[objName];
+    return isSelected && h('.bg-white', { style: 'height: 20em' }, remoteObject?.match({
+      NotAsked: () => null,
+      Loading: () =>
+        h('.absolute-fill.flex-column.items-center.justify-center.f5', [spinner(5), h('', 'Loading Objects')]),
+      Success: (data) => draw(data),
+      Failure: () => null, // Notification is displayed
+    }));
   }
-  return;
+  return null;
 };
