+++ QualityControl/public/object/objectTreePage.js
@@ -14,7 +14,7 @@
 
 import { h, iconBarChart, iconCaretRight, iconResizeBoth, iconCaretBottom, iconCircleX } from '/js/src/index.js';
 import { spinner } from '../common/spinner.js';
-import { draw } from './objectDraw.js';
+import { draw } from '../common/object/draw.js';
 import timestampSelectForm from './../common/timestampSelectForm.js';
 import virtualTable from './virtualTable.js';
 import { qcObjectInfoPanel } from '../common/object/objectInfoCard.js';
@@ -94,7 +94,6 @@ const drawPlot = (model, object) => {
   const href = validFrom ?
     `?page=objectView&objectName=${name}&ts=${validFrom}&id=${id}`
     : `?page=objectView&objectName=${name}`;
-  const info = object;
   return h('', { style: 'height:100%; display: flex; flex-direction: column' }, [
     h('.item-action-row.flex-row.g1.p1', [
       downloadButton({
@@ -119,10 +118,10 @@ const drawPlot = (model, object) => {
         iconCircleX(),
       ),
     ]),
-    h('', { style: 'height:77%;' }, draw(model, name, { stat: true })),
+    h('', { style: 'height:77%;' }, draw(object, { stat: true })),
     h('.scroll-y', {}, [
       h('.w-100.flex-row', { style: 'justify-content: center' }, h('.w-80', timestampSelectForm(model))),
-      qcObjectInfoPanel(info, { 'font-size': '.875rem;' }),
+      qcObjectInfoPanel(object, { 'font-size': '.875rem;' }),
     ]),
   ]);
 };
